extract eta panel from model tab
